move scripts to src root
