+++ src/application/redux/containers/settings-show-mnemonic.container.ts
@@ -6,7 +6,6 @@ import SettingsShowMnemonicView, {
 
 const mapStateToProps = (state: RootReducerState): SettingsShowMnemonicProps => ({
   wallet: state.wallet,
-  network: state.app.network,
 });
 
 const SettingsShowMnemonic = connect(mapStateToProps)(SettingsShowMnemonicView);
