+++ src/popup.html
@@ -1,0 +1,21 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+  <!-- REMOVE IN PROD -->
+  <!--<script src="http://localhost:8097"></script>-->
+  <meta charset="utf-8" />
+  <meta name="viewport" content="width=device-width, initial-scale=1, user-scalable=no" />
+  <title>Marina</title>
+  <link rel="stylesheet" href="../src/presentation/styles/index.css" />
+  <link rel="stylesheet" href="../src/presentation/styles/fonts.css" />
+  <link rel="stylesheet" href="../src/presentation/styles/popup.css" />
+</head>
+
+<body class="text-grayDark">
+  <noscript>You need to enable JavaScript to run this app.</noscript>
+  <div id="root"></div>
+  <script src="../src/presentation/index.tsx"></script>
+</body>
+
+</html>
